backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
fdb==2.0.2
python-dotenv==1.0.0
httpx==0.25.2
apscheduler==3.10.4